requirements.txt
Flask>=2.0
Flask-SQLAlchemy
psycopg2-binary
flask-bcrypt
flask-login
python-dotenv